Move data generator subclass definitions into their own files
