--- integrations/lambda/src/lambda_sample/data_generator.py
@@ -1,2031 +1,0 @@
-import random
-from datetime import datetime, timedelta
-
-class DataGenerator:
-    def __init__(self, measure_templates: list, dimension_templates: list):
-        """
-        :param measure_templates: list: A list of dicts used to define the format of measures.
-            Format:
-            "name": str. Required. The name of the measure.
-            "type": str. Optional. The Timestream for LiveAnalytics data type of the measure. Valid options are "DOUBLE",
-                "BIGINT", "BOOLEAN", and "VARCHAR". Defaults to "DOUBLE".
-            "max_variation": Optional. The maximum amount a measure value can changed over time, positively or negatively.
-                For example, with a value of 5.0, measure values will increment by a max of 5.0 and a min of -5.0. Defaults to 1.5.
-            "max": Optional. The maximum measure value. Defaults to 100.0.
-            "min": Optional. The minimum measure value. Defaults to 0.0.
-            "random_options": Optional. A list of values the measure value can have. All elements of the list should be the same data type and
-                match the data type specified by the "type" field. This field overrides "max_variation", "max", and "min".
-                Values will be reused, in the same way as the dimension_templates "unique_options" field.
-            
-            Example:
-            [
-                {
-                    "name": "temperature_celsius",
-                    "type": "DOUBLE",
-                    "max_variation": 2.0,
-                    "max": 40.0,
-                    "min": 30.0
-                },
-                {
-                    "name": "symptoms",
-                    "type": "VARCHAR",
-                    "random_options": ["none", "headache", "shortness of breath", "fatigue", "nausea"]
-                }
-            ]
-
-        :param dimension_templates: list: A list of dicts used to define the format of dimensions.
-            Format:
-            "name": str. Required. The name of the dimension.
-            "value_length": int. Optional. The length of the dimension value, when it is randomly generated. If neither this nor
-                "options" are provided, defaults to 10.
-            "random_options": [str]. Optional. An array of strings to pick at random as options for the dimension value.
-                Has precedence over "value_length". Values will be reused.
-            "unique_options": [str]. Optional. An array of strings to pick serially for the dimension value. Each value in
-                this array will be used once.
-
-            Example:
-            [
-                {
-                    "name": "device_id",
-                    "value_length": 9
-                },
-                {
-                    "name": "region",
-                    "random_options": ["us-east-1", "us-west-2"]
-                }
-            ]
-        """
-        self.measure_templates = measure_templates
-        self.dimension_templates = dimension_templates
-
-    def generate(self, start_date: datetime, end_date: datetime, reporting_frequency: timedelta,
-                num_entities: int, precision="MILLISECONDS", generate_unique_options_fallback=False) -> list:
-        """
-        Generates time series data.
-
-        :param start_date: datetime: The start date to use when generating records. This cannot be older in hours
-            than the memory retention period in hours value for the table.
-        :param end_date: datetime: The end date to use when generating records. The maximum end date
-            Timestream for LiveAnalytics allows is 15 minutes in the future.
-        :param reporting_frequency: timedelta: The frequency that records are generated by all entities, for example,
-            every 2 seconds, every 5 hours, etc.
-        :param num_entities: int: The number of entities that will report for each timestamp, for example,
-            the number of servers or number of stocks.
-        :param precision: str: The precision to use for record timestamps. Valid options are "MILLISECONDS",
-            "SECONDS", and "MICROSECONDS". (Default value = "MILLISECONDS")
-        :param generate_unique_options_fallback: bool: Whether to generate random strings for dimension values
-            after all values in a dimension template's "unique_options" array have been used. (Default value = False)
-        :returns: A list containing Timestream for LiveAnalytics measures and dimensions.
-        """
-
-        # Construct entities (e.g., servers, weather reporting stations, stocks, etc.)
-        entities = []
-        for _ in range(num_entities):
-            entity = {"latest_measures": {}}
-            for dimension_template in self.dimension_templates:
-                dimension_value_length = 20
-
-                if "value_length" in dimension_template:
-                    dimension_value_length = dimension_template["value_length"]
-
-                # Unique options that should not be reused. Stock symbols are
-                # an example of this.
-                if "unique_options" in dimension_template:
-                    if len(dimension_template["unique_options"]) > 0:
-                        dimension_value = dimension_template["unique_options"][0]
-                        # Dimensions must be unique. Each time a choice is chosen, remove it from the list.
-                        dimension_template["unique_options"].pop(0)
-                    elif generate_unique_options_fallback:
-                        # Generate a fallback value, since we've run out of options and the user
-                        # has specified that they want more unique values generated.
-                        dimension_value = self._generate_random_string(dimension_value_length)
-                
-                # Options that are reused. Server regions are an example of this.
-                elif "random_options" in dimension_template:
-                    dimension_value = random.choice(dimension_template["random_options"])
-
-                elif "unique_options" not in dimension_template and "random_options" not in dimension_template:
-                    dimension_value = self._generate_random_string(dimension_value_length)
-
-                entity[dimension_template["name"]] = dimension_value
-                    
-            entities.append(entity)
-
-        records = []
-        current_date = start_date
-        while current_date <= end_date:
-            # Each entity has a record for a single timestamp
-            for entity in entities:
-                dimensions = []
-                for key in entity:
-                    if key != "latest_measures":
-                        dimension = {
-                            "Name": key,
-                            "Value": entity[key],
-                            "DimensionValueType": "VARCHAR" # Not configurable
-                        }
-                        dimensions.append(dimension)
-
-                measures = []
-
-                for measure_template in self.measure_templates:
-                    if "name" not in measure_template:
-                        raise Exception(f"Measure template was missing name: {measure_template}")
-                    measure_name = measure_template["name"]
-
-                    # Optional template fields
-                    measure_value_type = "DOUBLE"
-                    if "type" in measure_template:
-                        measure_value_type = str(measure_template["type"]).strip().upper()
-                    max_variation = 1.5
-                    if "max_variation" in measure_template:
-                        max_variation = measure_template["max_variation"]
-                    max_value = 100.0
-                    if "max" in measure_template:
-                        max_value = measure_template["max"]
-                    min_value = 0.0
-                    if "min" in measure_template:
-                        min_value = measure_template["min"]
-                    varchar_length = 10
-                    if "varchar_length" in measure_template:
-                        varchar_length = measure_template["varchar_length"]
-
-                    measure = {
-                        "MeasureName": measure_name,
-                        "MeasureValueType": measure_value_type
-                    }
-
-                    measure_value = None
-                    if "random_options" in measure_template:
-                        measure_value = random.choice(measure_template["random_options"])
-                    else:
-                        if current_date == start_date:
-                            if measure_value_type == "DOUBLE":
-                                measure_value = random.uniform(min_value, max_value)
-                            elif measure_value_type == "VARCHAR":
-                                measure_value = self._generate_random_string(varchar_length)
-                            elif measure_value_type == "BIGINT":
-                                measure_value = int(random.uniform(min_value, max_value))
-                            elif measure_value_type == "BOOLEAN":
-                                measure_value = random.choice([True, False])
-                            else:
-                                raise Exception("Measure value type not recognized")
-                        else:
-                            if measure_value_type == "DOUBLE":
-                                measure_value = max(min_value, min(entity["latest_measures"][measure_name] + random.uniform(-max_variation, max_variation), max_value))
-                            elif measure_value_type == "VARCHAR":
-                                measure_value = self._generate_random_string(varchar_length)
-                            elif measure_value_type == "BIGINT":
-                                measure_value = int(max(min_value, min(entity["latest_measures"][measure_name] + int(random.uniform(-max_variation, max_variation)), max_value)))
-                            elif measure_value_type == "BOOLEAN":
-                                measure_value = random.choice([True, False])
-                            else:
-                                raise Exception("Measure value type not recognized")
-                        
-                    # Store the actual value in the entity for future iteration
-                    entity["latest_measures"][measure_name] = measure_value
-
-                    # Timestream requires that all data be inserted as a string
-                    measure["MeasureValue"] = str(measure_value)
-
-                    measures.append(measure)
-                
-                precision = precision.strip().upper()
-                if precision == "SECONDS":
-                    timestamp = str(int(current_date.timestamp()))
-                if precision == "MICROSECONDS":
-                    timestamp = str(int(current_date.timestamp() * 1_000_000))
-                else:
-                    # Default to millisecond precision
-                    timestamp = str(int(current_date.timestamp() * 1_000))
-
-                record = {
-                    "Dimensions": dimensions,
-                    "Time": timestamp,
-                    "Measures": measures
-                }
-                records.append(record)
-            current_date += reporting_frequency
-        return records
-    
-    def _generate_random_string(self, length: int) -> str:
-        """
-        Generates a random alphanumeric string.
-
-        :param length: The length of the string to generate.
-        :returns: A random string.
-        """
-
-        letters = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"
-        return ''.join(random.choice(letters) for _ in range(length))
-    
-    def generate_dashboard(self, grafana_data_source_name: str, database_name: str, table_name: str) -> dict:
-        """
-        Generates a Grafana dashboard.
-
-        :param grafana_data_source_name: str: The name of an existing Timestream for LiveAnalytics data source in Grafana.
-        :param database_name: str: The Timestream for LiveAnalytics database the panels will use.
-        :param table_name: str: The Timestream for LiveAnalytics table the panels will use.
-        :returns: A Grafana dashboard as a dict.
-        """
-
-        # Each measure will have a panel
-        panels = []
-        for _, measure_template in enumerate(self.measure_templates):
-            measure_name = measure_template['name']
-            query = ""
-            if len(self.measure_templates) > 1:
-                dimensions = ', '.join(f"\"{dimension_template['name']}\"" for dimension_template in self.dimension_templates)
-                query = f'SELECT time, "{measure_name}", {dimensions} FROM "{database_name}"."{table_name}" ORDER BY time ASC'
-            elif len(self.measure_templates) == 1:
-                query = f'SELECT * FROM "{database_name}"."{table_name}" ORDER BY time ASC'
-            
-            panel = {
-                "datasource": grafana_data_source_name,
-                "gridPos": {
-                    "h": 22,
-                    "w": 20,
-                    "x": 0,
-                    "y": 0
-                },
-                "targets": [
-                    {
-                        "datasource": grafana_data_source_name,
-                        "rawQuery": query,
-                    }
-                ],
-                "title": f"{measure_name}",
-                "type": "timeseries"
-            }
-            panels.append(panel)
-
-        dashboard = {
-            "panels": panels,
-            "title": "Amazon Timestream for LiveAnalytics Sample Dashboard"
-        }
-
-        return dashboard
-
-# DataGenerator subclass definitions
-
-class DevOpsDataGenerator(DataGenerator):
-    def __init__(self):
-        self.measure_templates = [
-            {
-                "name": "cpu_usage",
-                "type": "DOUBLE",
-                "max_variation": 1.2,
-                "max": 100.0,
-                "min": 0.0
-            },
-            {
-                "name": "mem_usage",
-                "type": "DOUBLE",
-                "max_variation": 2.3,
-                "max": 100.0,
-                "min": 0.0
-            },
-            {
-                "name": "disk_usage",
-                "type": "DOUBLE",
-                "max_variation": 0.5,
-                "max": 100.0,
-                "min": 0.0
-            },
-            {
-                "name": "network_in",
-                "type": "DOUBLE",
-                "max_variation": 100,
-                "max": 5000,
-                "min": 0
-            },
-            {
-                "name": "network_out",
-                "type": "DOUBLE",
-                "max_variation": 20,
-                "max": 2000,
-                "min": 0
-            }
-        ]
-        self.dimension_templates = [
-            {
-                "name": "server_id",
-                "value_length": 14
-            },
-            {
-                "name": "region",
-                "random_options": ["ca-central-1", "ca-west-1", "us-east-1", "us-east-2", "us-west-1", "us-west-2", "sa-east-1", "eu-central-1", "eu-west-1", "eu-west-2", "eu-south-1", "eu-west-3"]
-            }
-        ]
-    
-    def generate_dashboard(self, grafana_data_source_name, database_name, table_name) -> dict:
-        return {
-            "panels": [
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "unit": "percent"
-                        },
-                    },
-                    "gridPos": {
-                        "h": 16,
-                        "w": 9,
-                        "x": 0,
-                        "y": 0
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, server_id, cpu_usage AS \"CPU Usage\", mem_usage as \"Memory Usage\" FROM \"{database_name}\".\"{table_name}\" WHERE server_id = '$Server' ORDER BY time ASC",
-                        }
-                    ],
-                    "title": "CPU / Memory Usage",
-                    "type": "timeseries"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "color": {
-                                "mode": "continuous-BlPu"
-                            },
-                            "unit": "binBps"
-                        },
-                    },
-                    "gridPos": {
-                        "h": 16,
-                        "w": 10,
-                        "x": 9,
-                        "y": 0
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, network_in AS \"Network In\", network_out AS \"Network Out\", server_id FROM \"{database_name}\".\"{table_name}\" WHERE server_id = '$Server' ORDER BY time ASC",
-                        }
-                    ],
-                    "title": "Network In / Network Out",
-                    "type": "timeseries"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "unit": "percent"
-                        },
-                    },
-                    "gridPos": {
-                        "h": 8,
-                        "w": 5,
-                        "x": 0,
-                        "y": 16
-                    },
-                    "options": {
-                        "reduceOptions": {
-                            "calcs": [
-                                "mean"
-                            ],
-                        },
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, AVG(disk_usage) FROM \"{database_name}\".\"{table_name}\" WHERE server_id = '$Server' GROUP BY time",
-                        }
-                    ],
-                    "title": "Average Disk Usage Percentage",
-                    "type": "stat"
-                }
-            ],
-            "templating": {
-                "list": [
-                    {
-                        "definition": f"SELECT DISTINCT server_id FROM \"{database_name}\".\"{table_name}\"",
-                        "name": "Server",
-                        "query": f"SELECT DISTINCT server_id FROM \"{database_name}\".\"{table_name}\"",
-                        "refresh": 1,
-                        "sort": 1,
-                        "type": "query"
-                    }
-                ]
-            },
-            "time": {
-                "from": "now-15m",
-                "to": "now"
-            },
-            "title": "Amazon Timestream for LiveAnalytics Sample Dashboard"
-        }
-
-class IoTDataGenerator(DataGenerator):
-    def __init__(self):
-        self.measure_templates = [
-            {
-                "name": "temperature_celsius",
-                "type": "DOUBLE",
-                "max_variation": 0.5,
-                "max": 60,
-                "min": -30
-            },
-            {
-                "name": "relative_humidity",
-                "type": "DOUBLE",
-                "max_variation": 0.3,
-                "max": 100.0,
-                "min": 0.0
-            },
-            {
-                "name": "battery_level",
-                "type": "BIGINT",
-                "max_variation": 1,
-                "max": 100,
-                "min": 1 # All devices have enough battery to report
-            },
-            {
-                "name": "velocity",
-                "type": "DOUBLE",
-                "max_variation": 4.2,
-                "max": 100.0,
-                "min": 0.0
-            }
-        ]
-        self.dimension_templates = [
-            {
-                "name": "device_id",
-                "value_length": 14
-            }
-        ]
-
-    def generate_dashboard(self, grafana_data_source_name, database_name, table_name) -> dict:
-        return {
-            "panels": [
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "color": {
-                                "mode": "thresholds"
-                            },
-                            "mappings": [],
-                            "thresholds": {
-                                "mode": "absolute",
-                                "steps": [
-                                    {
-                                        "color": "red",
-                                        "value": None
-                                    },
-                                    {
-                                        "color": "yellow",
-                                        "value": 15
-                                    },
-                                    {
-                                        "color": "green",
-                                        "value": 40
-                                    }
-                                ]
-                            },
-                            "unit": "percent"
-                        },
-                    },
-                    "gridPos": {
-                        "h": 8,
-                        "w": 12,
-                        "x": 0,
-                        "y": 0
-                    },
-                    "options": {
-                        "reduceOptions": {
-                            "values": True
-                        }
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT device_id, battery_level FROM \"{database_name}\".\"{table_name}\" WHERE time = (SELECT MAX(time) FROM \"{database_name}\".\"{table_name}\" AS sub WHERE sub.device_id = \"{database_name}\".\"{table_name}\".device_id)",
-                        }
-                    ],
-                    "title": "Current Device Battery Level",
-                    "type": "gauge"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "gridPos": {
-                        "h": 8,
-                        "w": 12,
-                        "x": 12,
-                        "y": 0
-                    },
-                    "options": {
-                        "cellValues": {
-                            "unit": "velocityms"
-                        },
-                        "color": {
-                            "mode": "scheme",
-                            "scheme": "RdYlBu",
-                        },
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "format": 1,
-                            "rawQuery": f"SELECT time, device_id, velocity FROM \"{database_name}\".\"{table_name}\" ORDER BY time ASC",
-                        }
-                    ],
-                    "title": "Device Velocity",
-                    "type": "heatmap"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "thresholds": {
-                                "mode": "absolute",
-                                "steps": [
-                                {
-                                    "color": "green",
-                                    "value": None
-                                },
-                                {
-                                    "color": "#EAB839",
-                                    "value": 50
-                                },
-                                {
-                                    "color": "red",
-                                    "value": 70
-                                }
-                                ]
-                            },
-                            "unit": "celsius"
-                        },
-                    },
-                    "gridPos": {
-                        "h": 8,
-                        "w": 12,
-                        "x": 0,
-                        "y": 8
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "format": 1,
-                            "rawQuery": f"SELECT time, device_id, temperature_celsius FROM \"{database_name}\".\"{table_name}\" ORDER BY time ASC",
-                        }
-                    ],
-                    "title": "Device Temperature",
-                    "type": "timeseries"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "thresholds": {
-                                "mode": "absolute",
-                                "steps": [
-                                    {
-                                        "color": "green",
-                                        "value": None
-                                    },
-                                    {
-                                        "color": "red",
-                                        "value": 80
-                                    }
-                                ]
-                            },
-                            "unit": "humidity"
-                        },
-                    },
-                    "gridPos": {
-                        "h": 8,
-                        "w": 12,
-                        "x": 12,
-                        "y": 8
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "format": 1,
-                            "rawQuery": f"SELECT time, device_id, relative_humidity FROM \"{database_name}\".\"{table_name}\" ORDER BY time ASC",
-                        }
-                    ],
-                    "title": "Device Relative Humidity",
-                    "type": "timeseries"
-                }
-            ],
-            "time": {
-                "from": "now-15m",
-                "to": "now"
-            },
-            "title": "Amazon Timestream for LiveAnalytics Sample Dashboard"
-        }
-
-class StockMarketDataGenerator(DataGenerator):
-    def __init__(self):
-        self.measure_templates = [
-            {
-                "name": "volume",
-                "type": "BIGINT",
-                "max_variation": 30000,
-                "max": 100000000,
-                "min": 1
-            },
-            {
-                "name": "market_cap",
-                "type": "BIGINT",
-                "max_variation": 100,
-                "max": 100000000000,
-                "min": 1000000000,
-            },
-            {
-                "name": "price_change",
-                "type": "DOUBLE",
-                "max_variation": 0.5,
-                "max": 1000.0,
-                "min": -1000.0
-            },
-            {
-                "name": "percentage_change",
-                "type": "DOUBLE",
-                "max_variation": 10.0,
-                "max": 100.0,
-                "min": -100.0
-            }
-        ]
-        self.dimension_templates = [
-            {
-                "name": "stock_symbol",
-                "unique_options": ["AAPL", "TSLA", "DJIA", "SPOT", "NFLX", "MSFT", "MCD", "PG", "KO", "MMM", "IBM", "AMZN", "VZ", "JNJ", "WMT"]
-            }
-        ]
-
-    def generate(self, start_date, end_date, reporting_frequency, num_entities, precision="MILLISECONDS", generate_unique_options_fallback=False) -> list:
-        num_stock_symbols = len(self.dimension_templates[0]["unique_options"])
-        if num_entities > num_stock_symbols and not generate_unique_options_fallback:
-            raise Exception(f"num_entities ({num_entities}) was greater than the number of stock symbols ({num_stock_symbols})")
-        return super().generate(start_date, end_date, reporting_frequency, num_entities, precision, generate_unique_options_fallback)
-    
-    def generate_dashboard(self, grafana_data_source_name, database_name, table_name) -> dict:
-        return {
-            "panels": [
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "overrides": [
-                            {
-                                "matcher": {
-                                    "id": "byName",
-                                    "options": "Percentage Change"
-                                },
-                                "properties": [
-                                    {
-                                        "id": "custom.cellOptions",
-                                        "value": {
-                                            "type": "color-text"
-                                        }
-                                    },
-                                    {
-                                        "id": "thresholds",
-                                        "value": {
-                                            "mode": "absolute",
-                                            "steps": [
-                                                {
-                                                    "color": "red",
-                                                    "value": None
-                                                },
-                                                {
-                                                    "color": "green",
-                                                    "value": 0
-                                                }
-                                            ]
-                                        }
-                                    }
-                                ]
-                            },
-                            {
-                                "matcher": {
-                                    "id": "byName",
-                                    "options": "Market Cap"
-                                },
-                                "properties": [
-                                    {
-                                        "id": "unit",
-                                        "value": "none"
-                                    }
-                                ]
-                            }
-                        ]
-                    },
-                    "gridPos": {
-                        "h": 22,
-                        "w": 20,
-                        "x": 0,
-                        "y": 0
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT stock_symbol AS \"Name\", market_cap AS \"Market Cap\", percentage_change AS \"Percentage Change\", time AS \"Time\" FROM \"{database_name}\".\"{table_name}\" WHERE time = (SELECT MAX(time) FROM \"{database_name}\".\"{table_name}\" AS sub WHERE sub.stock_symbol = \"{database_name}\".\"{table_name}\".stock_symbol)",
-                        }
-                    ],
-                    "title": "Intraday Stock Market",
-                    "type": "table"
-                }
-            ],
-            "time": {
-                "from": "now-15m",
-                "to": "now"
-            },
-            "title": "Amazon Timestream for LiveAnalytics Sample Dashboard"
-        }
-
-class WeatherDataGenerator(DataGenerator):
-    def __init__(self):
-        self.measure_templates = [
-            {
-                "name": "temperature_celsius",
-                "type": "DOUBLE",
-                "max_variation": 1.5,
-                "max": 65.0,
-                "min": -30
-            },
-            {
-                "name": "relative_humidity",
-                "type": "DOUBLE",
-                "max_variation": 0.5,
-                "max": 100.0,
-                "min": 0.0
-            },
-            {
-                "name": "wind_speed_kph",
-                "type": "DOUBLE",
-                "max_variation": 5.5,
-                "max": 407.164,
-                "min": 0.0
-            },
-            {
-                "name": "precipitation_mm",
-                "type": "DOUBLE",
-                "max_variation": 1.5,
-                "max": 60.0,
-                "min": 0.0
-            },
-            {
-                "name": "cloud_percentage",
-                "type": "DOUBLE",
-                "max_variation": 10.0,
-                "max": 100.0,
-                "min": 0.0
-            },
-            {
-                "name": "pressure_hpa",
-                "type": "BIGINT",
-                "max_variation": 10,
-                "max": 1050,
-                "min": 950
-            },
-            {
-                "name": "visibility_km",
-                "type": "BIGINT",
-                "max_variation": 20,
-                "max": 200,
-                "min": 1
-            }
-        ]
-        self.dimension_templates = [
-            {
-                "name": "location", 
-                "unique_options": ["San Francisco, CA", "Chicago, IL", "New York, NY", "Miami, FL", "Dallas, TX", "Gary, IN", "Las Vegas, NV", "San Diego, CA", "Portland, OR", "Seattle, WA", "New Orleans, LA", "Fargo, ND", "Albuquerque, NM"]
-            }
-        ]
-
-    def generate(self, start_date, end_date, reporting_frequency, num_entities, precision="MILLISECONDS", generate_unique_options_fallback=False) -> list:
-        num_locations = len(self.dimension_templates[0]["unique_options"])
-        if num_entities > num_locations and not generate_unique_options_fallback:
-            raise Exception(f"num_entities ({num_entities}) was greater than the number of locations ({num_locations})")
-        return super().generate(start_date, end_date, reporting_frequency, num_entities, precision, generate_unique_options_fallback)
-    
-    def generate_dashboard(self, grafana_data_source_name, database_name, table_name) -> dict:
-        return {
-            "panels": [
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "color": {
-                                "mode": "thresholds"
-                            },
-                            "thresholds": {
-                                "mode": "absolute",
-                                "steps": [
-                                    {
-                                        "color": "green",
-                                        "value": None
-                                    },
-                                    {
-                                        "color": "red",
-                                        "value": 40
-                                    }
-                                ]
-                            },
-                            "unit": "velocitykmh"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 10,
-                        "w": 6,
-                        "x": 0,
-                        "y": 0
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, wind_speed_kph AS \"Wind Speed\", location FROM \"{database_name}\".\"{table_name}\" WHERE location = '$Location' ORDER BY time ASC",
-                        }
-                    ],
-                    "title": "Wind Speed",
-                    "type": "stat"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "color": {
-                                "mode": "thresholds"
-                            },
-                            "thresholds": {
-                                "mode": "absolute",
-                                "steps": [
-                                    {
-                                        "color": "green",
-                                        "value": None
-                                    },
-                                    {
-                                        "color": "red",
-                                        "value": 80
-                                    }
-                                ]
-                            },
-                            "unit": "humidity"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 10,
-                        "w": 6,
-                        "x": 6,
-                        "y": 0
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, relative_humidity AS \"Relative Humidity\", location FROM \"{database_name}\".\"{table_name}\" WHERE location = '$Location' ORDER BY time ASC",
-                        }
-                    ],
-                    "title": "Relative Humidity",
-                    "type": "stat"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "color": {
-                                "fixedColor": "#00d7ff",
-                                "mode": "fixed"
-                            },
-                            "unit": "pressurehpa"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 10,
-                        "w": 6,
-                        "x": 12,
-                        "y": 0
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, pressure_hpa AS \"Pressure\", location FROM \"{database_name}\".\"{table_name}\" WHERE location = '$Location' ORDER BY time ASC",
-                        }
-                    ],
-                    "title": "Pressure",
-                    "type": "stat"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "color": {
-                                "mode": "thresholds"
-                            },
-                            "custom": {
-                                "fillOpacity": 9
-                            },
-                            "thresholds": {
-                                "mode": "absolute",
-                                "steps": [
-                                    {
-                                        "color": "blue",
-                                        "value": None
-                                    },
-                                    {
-                                        "color": "green",
-                                        "value": 0
-                                    },
-                                    {
-                                        "color": "red",
-                                        "value": 30
-                                    },
-                                    {
-                                        "color": "dark-red",
-                                        "value": 50
-                                    }
-                                ]
-                            },
-                            "unit": "celsius"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 16,
-                        "w": 11,
-                        "x": 0,
-                        "y": 10
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, temperature_celsius AS \"Temperature\", location FROM \"{database_name}\".\"{table_name}\" WHERE location = '$Location' ORDER BY time ASC",
-                        }
-                    ],
-                    "title": "Temperature",
-                    "type": "timeseries"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "color": {
-                                "mode": "continuous-BlPu"
-                            },
-                            "custom": {
-                                "fillOpacity": 9
-                            },
-                            "thresholds": {
-                                "mode": "absolute",
-                                "steps": [
-                                    {
-                                        "color": "green",
-                                        "value": None
-                                    },
-                                    {
-                                        "color": "red",
-                                        "value": 45
-                                    }
-                                ]
-                            },
-                            "unit": "lengthmm"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 16,
-                        "w": 11,
-                        "x": 11,
-                        "y": 10
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, precipitation_mm AS \"Precipitation\", location FROM \"{database_name}\".\"{table_name}\" WHERE location = '$Location' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Precipitation",
-                    "type": "timeseries"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "color": {
-                                "mode": "continuous-BlPu"
-                            },
-                            "max": 100,
-                            "unit": "percent"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 17,
-                        "w": 11,
-                        "x": 0,
-                        "y": 26
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, cloud_percentage AS \"Cloud Percentage\", location FROM \"{database_name}\".\"{table_name}\" WHERE location = '$Location' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Cloud Coverage",
-                    "type": "timeseries"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "unit": "lengthkm"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 17,
-                        "w": 11,
-                        "x": 11,
-                        "y": 26
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, visibility_km as \"Visibility\", location FROM \"{database_name}\".\"{table_name}\" WHERE location = '$Location' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Visibility",
-                    "type": "timeseries"
-                }
-            ],
-            "templating": {
-                "list": [
-                    {
-                        "datasource": grafana_data_source_name,
-                        "definition": f"SELECT DISTINCT location FROM \"{database_name}\".\"{table_name}\"",
-                        "name": "Location",
-                        "query": f"SELECT DISTINCT location FROM \"{database_name}\".\"{table_name}\"",
-                        "refresh": 1,
-                        "sort": 1,
-                        "type": "query"
-                    }
-                ]
-            },
-            "time": {
-                "from": "now-15m",
-                "to": "now"
-            },
-            "title": "Amazon Timestream for LiveAnalytics Sample Dashboard"
-        }
-
-class GamingDataGenerator(DataGenerator):
-    def __init__(self):
-        self.measure_templates = [
-            {
-                "name": "X",
-                "type": "DOUBLE",
-                "max_variation": 2.1,
-                "max": 5000.0,
-                "min": -5000.0
-            },
-            {
-                "name": "Y",
-                "type": "DOUBLE",
-                "max_variation": 2.1,
-                "max": 5000.0,
-                "min": -5000.0
-            },
-            {
-                "name": "Z",
-                "type": "DOUBLE",
-                "max_variation": 2.1,
-                "max": 5000.0,
-                "min": -5000.0
-            },
-            {
-                "name": "health",
-                "type": "BIGINT",
-                "max_variation": 60,
-                "max": 100,
-                "min": 1
-            },
-            {
-                "name": "ping",
-                "type": "BIGINT",
-                "max_variation": 10,
-                "max": 250,
-                "min": 25
-            },
-            {
-                "name": "current_equip_value",
-                "type": "BIGINT",
-                "max_variation": 150,
-                "max": 1000000,
-                "min": 10
-            },
-            {
-                "name": "flash_duration",
-                "type": "DOUBLE",
-                "max_variation": 0.3,
-                "max": 10.0,
-                "min": 0.0
-            },
-            {
-                "name": "pitch",
-                "type": "DOUBLE",
-                "max_variation": 20.0,
-                "max": 90.0,
-                "min": -90.0
-            },
-            {
-                "name": "yaw",
-                "type": "DOUBLE",
-                "max_variation": 0.8,
-                "max": 360.0,
-                "min": 0.0
-            }
-        ]
-        self.dimension_templates = [
-            {
-                "name": "player_id",
-                "value_length": 25
-            },
-            {
-                "name": "player_name",
-                "value_length": 15
-            },
-            {
-                "name": "clan",
-                "random_options": ["mosdeff", "green_berets", "golden_ducks", "roberts", "club_z"]
-            }
-        ]
-
-    def generate_dashboard(self, grafana_data_source_name, database_name, table_name) -> dict:
-        return {
-            "panels": [
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "color": {
-                                "mode": "thresholds"
-                            },
-                            "thresholds": {
-                                "mode": "absolute",
-                                "steps": [
-                                    {
-                                        "color": "red",
-                                        "value": None
-                                    },
-                                    {
-                                        "color": "green",
-                                        "value": 35
-                                    }
-                                ]
-                            }
-                        }
-                    },
-                    "gridPos": {
-                        "h": 8,
-                        "w": 4,
-                        "x": 0,
-                        "y": 0
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, health, player_id FROM \"{database_name}\".\"{table_name}\" WHERE player_id = '$Player' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Health",
-                    "type": "stat"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "unit": "currencyUSD"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 8,
-                        "w": 4,
-                        "x": 4,
-                        "y": 0
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, current_equip_value, player_id FROM \"{database_name}\".\"{table_name}\" WHERE player_id = '$Player' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Equip Value",
-                    "type": "stat"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "color": {
-                                "fixedColor": "purple",
-                                "mode": "shades"
-                            },
-                            "unit": "ms"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 8,
-                        "w": 7,
-                        "x": 8,
-                        "y": 0
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, ping, player_id FROM \"{database_name}\".\"{table_name}\" WHERE player_id = '$Player' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Ping",
-                    "type": "timeseries"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "custom": {
-                                "show": "lines"
-                            }
-                        }
-                    },
-                    "gridPos": {
-                        "h": 20,
-                        "w": 8,
-                        "x": 0,
-                        "y": 8
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, X, Z, player_id FROM \"{database_name}\".\"{table_name}\" WHERE player_id = '$Player' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Coordinates",
-                    "type": "xychart"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "gridPos": {
-                        "h": 19,
-                        "w": 7,
-                        "x": 8,
-                        "y": 8
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, Y, player_id FROM \"{database_name}\".\"{table_name}\" WHERE player_id = '$Player' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Elevation",
-                    "type": "timeseries"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "gridPos": {
-                        "h": 19,
-                        "w": 8,
-                        "x": 0,
-                        "y": 28
-                    },
-                    "options": {
-                        "cellValues": {
-                            "unit": "s"
-                        },
-                        "color": {
-                            "mode": "scheme",
-                            "scheme": "RdYlGn"
-                        }
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, flash_duration, player_id FROM \"{database_name}\".\"{table_name}\" WHERE player_id = '$Player' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Flash Duration",
-                    "type": "heatmap"
-                }
-            ],
-            "templating": {
-                "list": [
-                    {
-                        "datasource": grafana_data_source_name,
-                        "definition": f"SELECT DISTINCT player_id FROM \"{database_name}\".\"{table_name}\"",
-                        "name": "Player",
-                        "query": f"SELECT DISTINCT player_id FROM \"{database_name}\".\"{table_name}\"",
-                        "refresh": 1,
-                        "sort": 1,
-                        "type": "query"
-                    }
-                ]
-            },
-            "time": {
-                "from": "now-15m",
-                "to": "now"
-            },
-            "title": "Amazon Timestream for LiveAnalytics Sample Dashboard"
-        }
-
-class AirQualityDataGenerator(DataGenerator):
-    def __init__(self):
-        self.measure_templates = [
-            {
-                "name": "PM2.5",
-                "type": "DOUBLE",
-                "max_variation": 1.0,
-                "max": 150.0,
-                "min": 15.0
-            },
-            {
-                "name": "PM10",
-                "type": "DOUBLE",
-                "max_variation": 1.0,
-                "max": 150.0,
-                "min": 15.0
-            },
-            {
-                "name": "CO_ppm",
-                "type": "DOUBLE",
-                "max_variation": 5.0,
-                "max": 100.0,
-                "min": 0.1
-            },
-            {
-                "name": "NO2_ppb",
-                "type": "DOUBLE",
-                "max_variation": 1.0,
-                "max": 300.0,
-                "min": 0.5
-            },
-            {
-                "name": "O2_percentage",
-                "type": "DOUBLE",
-                "max_variation": 1.0,
-                "max": 25.0,
-                "min": 20.8
-            }
-        ]
-        self.dimension_templates = [
-            {
-                "name": "city",
-                "unique_options": ["Los Angeles", "New York", "Vancouver", "Sydney", "Delhi", "Beijing", "London", "Miami", "Toronto", "Seattle", "Amsterdam"]
-            }
-        ]
-
-    def generate(self, start_date, end_date, reporting_frequency, num_entities, precision="MILLISECONDS", generate_unique_options_fallback=False) -> list:
-        num_cities = len(self.dimension_templates[0]["unique_options"])
-        if num_entities > num_cities and not generate_unique_options_fallback:
-            raise Exception(f"num_entities ({num_entities}) was greater than the number of cities ({num_cities})")
-        return super().generate(start_date, end_date, reporting_frequency, num_entities, precision, generate_unique_options_fallback)
-    
-    def generate_dashboard(self, grafana_data_source_name, database_name, table_name) -> dict:
-        return {
-            "panels": [
-                {
-                    "datasource": grafana_data_source_name,
-                    "gridPos": {
-                        "h": 22,
-                        "w": 20,
-                        "x": 0,
-                        "y": 0
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, \"PM2.5\", PM10, CO_ppm AS \"CO PPM\", NO2_ppb AS \"NO2 PPB\", O2_percentage AS \"O2 Percentage\", city FROM \"{database_name}\".\"{table_name}\" WHERE city = '$City' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Air Quality",
-                    "type": "timeseries"
-                }
-            ],
-            "templating": {
-                "list": [
-                    {
-                        f"definition": f"SELECT DISTINCT city FROM \"{database_name}\".\"{table_name}\"",
-                        "name": "City",
-                        "query": f"SELECT DISTINCT city FROM \"{database_name}\".\"{table_name}\"",
-                        "refresh": 1,
-                        "sort": 1,
-                        "type": "query"
-                    }
-                ]
-            },
-            "time": {
-                "from": "now-15m",
-                "to": "now"
-            },
-            "title": "Amazon Timestream for LiveAnalytics Sample Dashboard"
-        }
-
-class PatientDataGenerator(DataGenerator):
-    def __init__(self):
-        self.measure_templates = [
-            {
-                "name": "heart_rate_bpm",
-                "type": "BIGINT",
-                "max_variation": 5.0,
-                "max": 100,
-                "min": 60
-            },
-            {
-                "name": "oxygen_saturation_percentage",
-                "type": "DOUBLE",
-                "max_variation": 5.0,
-                "max": 100.0,
-                "min": 60.0
-            },
-            {
-                "name": "temperature_celsius",
-                "type": "DOUBLE",
-                "max_variation": 2.0,
-                "max": 40.0,
-                "min": 30.0
-            }
-        ]
-        self.dimension_templates = [
-            {
-                "name": "patient_id",
-                "value_length": 14
-            }
-        ]
-    
-    def generate_dashboard(self, grafana_data_source_name, database_name, table_name) -> dict:
-        return {
-            "panels": [
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "color": {
-                                "mode": "thresholds"
-                            },
-                            "thresholds": {
-                                "mode": "absolute",
-                                "steps": [
-                                    {
-                                        "color": "blue",
-                                        "value": None
-                                    },
-                                    {
-                                        "color": "green",
-                                        "value": 36.5
-                                    },
-                                    {
-                                        "color": "red",
-                                        "value": 37.5
-                                    }
-                                ]
-                            },
-                            "unit": "celsius"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 11,
-                        "w": 8,
-                        "x": 0,
-                        "y": 0
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, \"temperature_celsius\", \"patient_id\" FROM \"{database_name}\".\"{table_name}\" WHERE patient_id = '$Patient' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Body Temperature",
-                    "type": "gauge"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                        "color": {
-                            "mode": "thresholds"
-                        },
-                        "custom": {
-                            "fillOpacity": 7
-                        },
-                        "displayName": "BPM",
-                            "thresholds": {
-                                "mode": "absolute",
-                                "steps": [
-                                    {
-                                        "color": "green",
-                                        "value": None
-                                    },
-                                    {
-                                        "color": "#EAB839",
-                                        "value": 60
-                                    },
-                                    {
-                                        "color": "red",
-                                        "value": 80
-                                    }
-                                ]
-                            }
-                        }
-                    },
-                    "gridPos": {
-                        "h": 11,
-                        "w": 9,
-                        "x": 8,
-                        "y": 0
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, \"heart_rate_bpm\", \"patient_id\" FROM \"{database_name}\".\"{table_name}\" WHERE patient_id = '$Patient' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Heart Rate",
-                    "type": "timeseries"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "color": {
-                                "fixedColor": "blue",
-                                "mode": "shades"
-                            },
-                            "unit": "percent"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 13,
-                        "w": 17,
-                        "x": 0,
-                        "y": 11
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, \"oxygen_saturation_percentage\", \"patient_id\" FROM \"{database_name}\".\"{table_name}\" WHERE patient_id = '$Patient' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Oxygen Saturation",
-                    "type": "timeseries"
-                }
-            ],
-            "templating": {
-                "list": [
-                    {
-                        "datasource": grafana_data_source_name,
-                        "definition": f"SELECT DISTINCT patient_id FROM \"{database_name}\".\"{table_name}\"",
-                        "name": "Patient",
-                        "query": f"SELECT DISTINCT patient_id FROM \"{database_name}\".\"{table_name}\"",
-                        "refresh": 1,
-                        "sort": 1,
-                        "type": "query"
-                    }
-                ]
-            },
-            "time": {
-                "from": "now-15m",
-                "to": "now"
-            },
-            "title": "Amazon Timestream for LiveAnalytics Sample Dashboard"
-        }
-
-class EnergyDataGenerator(DataGenerator):
-    def __init__(self):
-        self.measure_templates = [
-            {
-                "name": "energy_usage_kWh",
-                "type": "BIGINT",
-                "max_variation": 50,
-                "max": 300,
-                "min": 10
-            },
-            {
-                "name": "occupancy",
-                "type": "BIGINT",
-                "max_variation": 20,
-                "max": 100,
-                "min": 0
-            },
-            {
-                "name": "temperature_celsius",
-                "type": "DOUBLE",
-                "max_variation": 0.5,
-                "max": 60,
-                "min": -30
-            },
-            {
-                "name": "relative_humidity",
-                "type": "DOUBLE",
-                "max_variation": 0.3,
-                "max": 100.0,
-                "min": 0.0
-            }
-        ]
-        self.dimension_templates = [
-            {
-                "name": "building_id",
-                "value_length": 14
-            }
-        ]
-    
-    def generate_dashboard(self, grafana_data_source_name, database_name, table_name) -> dict:
-        return {
-            "panels": [
-                {
-                    "datasource": grafana_data_source_name,
-                    "gridPos": {
-                        "h": 9,
-                        "w": 9,
-                        "x": 0,
-                        "y": 0
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "format": 1,
-                            "rawQuery": f"SELECT time, \"occupancy\", \"building_id\" FROM \"{database_name}\".\"{table_name}\" ORDER BY time ASC",
-                        }
-                    ],
-                    "title": "Building Occupancy",
-                    "type": "stat"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "gridPos": {
-                        "h": 22,
-                        "w": 18,
-                        "x": 0,
-                        "y": 9
-                    },
-                    "options": {
-                        "cellValues": {
-                            "unit": "kwatth"
-                        },
-                        "color": {
-                            "exponent": 0.5,
-                            "fill": "dark-orange",
-                            "mode": "scheme",
-                            "reverse": False,
-                            "scale": "exponential",
-                            "scheme": "RdYlGn",
-                            "steps": 64
-                        },
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "format": 1,
-                            "rawQuery": f"SELECT time, \"energy_usage_kWh\", \"building_id\" FROM \"{database_name}\".\"{table_name}\" ORDER BY time ASC",
-                        }
-                    ],
-                    "title": "Energy Usage",
-                    "type": "heatmap"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "color": {
-                                "mode": "thresholds"
-                            },
-                            "custom": {
-                                "fillOpacity": 6
-                            },
-                            "thresholds": {
-                                "mode": "absolute",
-                                "steps": [
-                                    {
-                                        "color": "blue",
-                                        "value": None
-                                    },
-                                    {
-                                        "color": "green",
-                                        "value": 20
-                                    },
-                                    {
-                                        "color": "red",
-                                        "value": 35
-                                    }
-                                ]
-                            },
-                            "unit": "celsius"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 17,
-                        "w": 18,
-                        "x": 0,
-                        "y": 31
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "format": 1,
-                            "rawQuery": f"SELECT time, \"temperature_celsius\", \"building_id\" FROM \"{database_name}\".\"{table_name}\" ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Building Temperature",
-                    "type": "timeseries"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "gridPos": {
-                        "h": 17,
-                        "w": 18,
-                        "x": 0,
-                        "y": 48
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "format": 1,
-                            "rawQuery": f"SELECT time, \"relative_humidity\", \"building_id\" FROM \"{database_name}\".\"{table_name}\" ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Building Relative Humidity",
-                    "type": "timeseries"
-                }
-            ],
-            "time": {
-                "from": "now-15m",
-                "to": "now"
-            },
-            "title": "Amazon Timestream for LiveAnalytics Sample Dashboard"
-        }
-
-class FlightDataGenerator(DataGenerator):
-    def __init__(self):
-        self.measure_templates = [
-            {
-                "name": "fuel_level_gallons",
-                "type": "BIGINT",
-                "max_variation": 40,
-                "max": 8000,
-                "min": 1
-            },
-            {
-                "name": "heading",
-                "type": "BIGINT",
-                "max_variation": 10,
-                "max": 360,
-                "min": 0
-            },
-            {
-                "name": "air_temperature_celsius",
-                "type": "DOUBLE",
-                "max_variation": 2.0,
-                "max": 40.0,
-                "min": -90.0
-            },
-            {
-                "name": "lat",
-                "type": "DOUBLE",
-                "max_variation": 0.5,
-                "max": 90.0,
-                "min": -90.0
-            },
-            {
-                "name": "lon",
-                "type": "DOUBLE",
-                "max_variation": 0.5,
-                "max": 180.0,
-                "min": -180.0
-            },
-            {
-                "name": "speed_knots",
-                "type": "BIGINT",
-                "max_variation": 10,
-                "max": 250,
-                "min": 200
-            }
-        ]
-        self.dimension_templates = [
-            {
-                "name": "flight_id",
-                "unique_options": ["FL123", "FL456", "FL890", "FL333", "FL100", "FL650", "FL256", "FL430", "FL211", "FL874"]
-            },
-            {
-                "name": "aircraft_type",
-                "random_options": ["Boeing 737", "Airbus A320"]
-            }
-        ]
-
-    def generate(self, start_date, end_date, reporting_frequency, num_entities, precision="MILLISECONDS", generate_unique_options_fallback=False) -> list:
-        num_flight_ids = len(self.dimension_templates[0]["unique_options"])
-        if num_entities > num_flight_ids and not generate_unique_options_fallback:
-            raise Exception(f"num_entities ({num_entities}) was greater than the number of flight IDs ({num_flight_ids})")
-        return super().generate(start_date, end_date, reporting_frequency, num_entities, precision, generate_unique_options_fallback)
-    
-    def generate_dashboard(self, grafana_data_source_name, database_name, table_name) -> dict:
-        return {
-            "panels": [
-                {
-                    "datasource": grafana_data_source_name,
-                    "gridPos": {
-                        "h": 22,
-                        "w": 20,
-                        "x": 0,
-                        "y": 0
-                    },
-                    "options": {
-                        "layers": [
-                            {
-                                "name": "Layer 1",
-                                "type": "route"
-                            }
-                        ],
-                        "view": {
-                            "id": "fit",
-                            "lat": 0,
-                            "lon": 0,
-                            "padding": 75,
-                            "zoom": 15
-                        }
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, \"lat\", \"lon\", \"flight_id\" FROM \"{database_name}\".\"{table_name}\" WHERE \"flight_id\" = '$Flight' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Flight Path",
-                    "type": "geomap"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "color": {
-                                "mode": "thresholds"
-                            },
-                            "thresholds": {
-                                "mode": "absolute",
-                                "steps": [
-                                    {
-                                        "color": "red",
-                                        "value": None
-                                    },
-                                    {
-                                        "color": "green",
-                                        "value": 100
-                                    }
-                                ]
-                            },
-                            "unit": "gallons"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 8,
-                        "w": 5,
-                        "x": 0,
-                        "y": 22
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, \"fuel_level_gallons\", \"flight_id\" FROM \"{database_name}\".\"{table_name}\" WHERE flight_id = '$Flight' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Fuel Level",
-                    "type": "stat"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "unit": "degree"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 8,
-                        "w": 5,
-                        "x": 5,
-                        "y": 22
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, \"heading\", \"flight_id\", \"aircraft_type\" FROM \"{database_name}\".\"{table_name}\" WHERE \"flight_id\" = '$Flight' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Heading",
-                    "type": "stat"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "unit": "celsius"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 8,
-                        "w": 5,
-                        "x": 10,
-                        "y": 22
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, \"air_temperature_celsius\", \"flight_id\" FROM \"{database_name}\".\"{table_name}\" WHERE \"flight_id\" = '$Flight' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Air Temperature",
-                    "type": "stat"
-                },
-                {
-                    "datasource": grafana_data_source_name,
-                    "fieldConfig": {
-                        "defaults": {
-                            "unit": "velocityknot"
-                        }
-                    },
-                    "gridPos": {
-                        "h": 8,
-                        "w": 5,
-                        "x": 15,
-                        "y": 22
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT time, \"speed_knots\", \"flight_id\" FROM \"{database_name}\".\"{table_name}\" WHERE \"flight_id\" = '$Flight' ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Speed",
-                    "type": "stat"
-                }
-            ],
-            "templating": {
-                "list": [
-                    {
-
-                        "definition": f"SELECT DISTINCT flight_id FROM \"{database_name}\".\"{table_name}\"",
-                        "name": "Flight",
-                        "query": f"SELECT DISTINCT flight_id FROM \"{database_name}\".\"{table_name}\"",
-                        "refresh": 1,
-                        "sort": 1,
-                        "type": "query"
-                    }
-                ]
-            },
-            "time": {
-                "from": "now-15m",
-                "to": "now"
-            },
-            "title": "Amazon Timestream for LiveAnalytics Sample Dashboard"
-        }
-
-class ExchangeRateDataGenerator(DataGenerator):
-    def __init__(self):
-        self.measure_templates = [
-            {
-                "name": "exchange_rate",
-                "type": "DOUBLE",
-                "max_variation": 1.0,
-                "max": 90.0,
-                "min": 0.62
-            }
-        ]
-        self.dimension_templates = [
-            {
-                "name": "currency_pair",
-                "unique_options": ["USD/EUR", "USD/CAD", "USD/GPP", "USD/CNY", "GBP/CAD", "GBP/JPY", "GBP/INR", "CHF/INR", "XAU/CNY", "UYU/CAD", "USD/XAU"]
-            }
-        ]
-
-    def generate(self, start_date, end_date, reporting_frequency, num_entities, precision="MILLISECONDS", generate_unique_options_fallback=False) -> list:
-        num_currency_pairs = len(self.dimension_templates[0]["unique_options"])
-        if num_entities > num_currency_pairs and not generate_unique_options_fallback:
-            raise Exception(f"num_entities ({num_entities}) was greater than the number of currency pairs ({num_currency_pairs})")
-        return super().generate(start_date, end_date, reporting_frequency, num_entities, precision, generate_unique_options_fallback)
-    
-    def generate_dashboard(self, grafana_data_source_name, database_name, table_name) -> dict:
-        return {
-            "panels": [
-                {
-                    "datasource": grafana_data_source_name,
-                    "gridPos": {
-                        "h": 22,
-                        "w": 20,
-                        "x": 0,
-                        "y": 0
-                    },
-                    "targets": [
-                        {
-                            "datasource": grafana_data_source_name,
-                            "rawQuery": f"SELECT currency_pair AS \"Currency Pair\", \"measure_value::double\" AS \"Exchange Rate\", time AS \"Time\" FROM \"{database_name}\".\"{table_name}\" WHERE time IN (SELECT MAX(time) FROM \"{database_name}\".\"{table_name}\") ORDER BY time ASC"
-                        }
-                    ],
-                    "title": "Currency Exchange Rate",
-                    "type": "table"
-                }
-            ],
-            "time": {
-                "from": "now-15m",
-                "to": "now"
-            },
-            "title": "Amazon Timestream for LiveAnalytics Sample Dashboard"
-        }
-
-class CustomDataGenerator(DataGenerator):
-    def __init__(self, measure_templates: list, dimension_templates: list):
-        self.measure_templates = measure_templates
-        self.dimension_templates = dimension_templates
