+++ integrations/lambda/src/lambda_sample/data_generator/custom_data_generator.py
@@ -1,0 +1,6 @@
+from .data_generator import DataGenerator
+
+class CustomDataGenerator(DataGenerator):
+    def __init__(self, measure_templates: list, dimension_templates: list):
+        self.measure_templates = measure_templates
+        self.dimension_templates = dimension_templates
