+++ integrations/lambda/src/lambda_sample/data_generator/data_generator.py
@@ -1,0 +1,267 @@
+import random
+from datetime import datetime, timedelta
+
+class DataGenerator:
+    def __init__(self, measure_templates: list, dimension_templates: list):
+        """
+        :param measure_templates: list: A list of dicts used to define the format of measures.
+            Format:
+            "name": str. Required. The name of the measure.
+            "type": str. Optional. The Timestream for LiveAnalytics data type of the measure. Valid options are "DOUBLE",
+                "BIGINT", "BOOLEAN", and "VARCHAR". Defaults to "DOUBLE".
+            "max_variation": Optional. The maximum amount a measure value can changed over time, positively or negatively.
+                For example, with a value of 5.0, measure values will increment by a max of 5.0 and a min of -5.0. Defaults to 1.5.
+            "max": Optional. The maximum measure value. Defaults to 100.0.
+            "min": Optional. The minimum measure value. Defaults to 0.0.
+            "random_options": Optional. A list of values the measure value can have. All elements of the list should be the same data type and
+                match the data type specified by the "type" field. This field overrides "max_variation", "max", and "min".
+                Values will be reused, in the same way as the dimension_templates "unique_options" field.
+            
+            Example:
+            [
+                {
+                    "name": "temperature_celsius",
+                    "type": "DOUBLE",
+                    "max_variation": 2.0,
+                    "max": 40.0,
+                    "min": 30.0
+                },
+                {
+                    "name": "symptoms",
+                    "type": "VARCHAR",
+                    "random_options": ["none", "headache", "shortness of breath", "fatigue", "nausea"]
+                }
+            ]
+
+        :param dimension_templates: list: A list of dicts used to define the format of dimensions.
+            Format:
+            "name": str. Required. The name of the dimension.
+            "value_length": int. Optional. The length of the dimension value, when it is randomly generated. If neither this nor
+                "options" are provided, defaults to 10.
+            "random_options": [str]. Optional. An array of strings to pick at random as options for the dimension value.
+                Has precedence over "value_length". Values will be reused.
+            "unique_options": [str]. Optional. An array of strings to pick serially for the dimension value. Each value in
+                this array will be used once.
+
+            Example:
+            [
+                {
+                    "name": "device_id",
+                    "value_length": 9
+                },
+                {
+                    "name": "region",
+                    "random_options": ["us-east-1", "us-west-2"]
+                }
+            ]
+        """
+        self.measure_templates = measure_templates
+        self.dimension_templates = dimension_templates
+
+    def generate(self, start_date: datetime, end_date: datetime, reporting_frequency: timedelta,
+                num_entities: int, precision="MILLISECONDS", generate_unique_options_fallback=False) -> list:
+        """
+        Generates time series data.
+
+        :param start_date: datetime: The start date to use when generating records. This cannot be older in hours
+            than the memory retention period in hours value for the table.
+        :param end_date: datetime: The end date to use when generating records. The maximum end date
+            Timestream for LiveAnalytics allows is 15 minutes in the future.
+        :param reporting_frequency: timedelta: The frequency that records are generated by all entities, for example,
+            every 2 seconds, every 5 hours, etc.
+        :param num_entities: int: The number of entities that will report for each timestamp, for example,
+            the number of servers or number of stocks.
+        :param precision: str: The precision to use for record timestamps. Valid options are "MILLISECONDS",
+            "SECONDS", and "MICROSECONDS". (Default value = "MILLISECONDS")
+        :param generate_unique_options_fallback: bool: Whether to generate random strings for dimension values
+            after all values in a dimension template's "unique_options" array have been used. (Default value = False)
+        :returns: A list containing Timestream for LiveAnalytics measures and dimensions.
+        """
+
+        # Construct entities (e.g., servers, weather reporting stations, stocks, etc.)
+        entities = []
+        for _ in range(num_entities):
+            entity = {"latest_measures": {}}
+            for dimension_template in self.dimension_templates:
+                dimension_value_length = 20
+
+                if "value_length" in dimension_template:
+                    dimension_value_length = dimension_template["value_length"]
+
+                # Unique options that should not be reused. Stock symbols are
+                # an example of this.
+                if "unique_options" in dimension_template:
+                    if len(dimension_template["unique_options"]) > 0:
+                        dimension_value = dimension_template["unique_options"][0]
+                        # Dimensions must be unique. Each time a choice is chosen, remove it from the list.
+                        dimension_template["unique_options"].pop(0)
+                    elif generate_unique_options_fallback:
+                        # Generate a fallback value, since we've run out of options and the user
+                        # has specified that they want more unique values generated.
+                        dimension_value = self._generate_random_string(dimension_value_length)
+                
+                # Options that are reused. Server regions are an example of this.
+                elif "random_options" in dimension_template:
+                    dimension_value = random.choice(dimension_template["random_options"])
+
+                elif "unique_options" not in dimension_template and "random_options" not in dimension_template:
+                    dimension_value = self._generate_random_string(dimension_value_length)
+
+                entity[dimension_template["name"]] = dimension_value
+                    
+            entities.append(entity)
+
+        records = []
+        current_date = start_date
+        while current_date <= end_date:
+            # Each entity has a record for a single timestamp
+            for entity in entities:
+                dimensions = []
+                for key in entity:
+                    if key != "latest_measures":
+                        dimension = {
+                            "Name": key,
+                            "Value": entity[key],
+                            "DimensionValueType": "VARCHAR" # Not configurable
+                        }
+                        dimensions.append(dimension)
+
+                measures = []
+
+                for measure_template in self.measure_templates:
+                    if "name" not in measure_template:
+                        raise Exception(f"Measure template was missing name: {measure_template}")
+                    measure_name = measure_template["name"]
+
+                    # Optional template fields
+                    measure_value_type = "DOUBLE"
+                    if "type" in measure_template:
+                        measure_value_type = str(measure_template["type"]).strip().upper()
+                    max_variation = 1.5
+                    if "max_variation" in measure_template:
+                        max_variation = measure_template["max_variation"]
+                    max_value = 100.0
+                    if "max" in measure_template:
+                        max_value = measure_template["max"]
+                    min_value = 0.0
+                    if "min" in measure_template:
+                        min_value = measure_template["min"]
+                    varchar_length = 10
+                    if "varchar_length" in measure_template:
+                        varchar_length = measure_template["varchar_length"]
+
+                    measure = {
+                        "MeasureName": measure_name,
+                        "MeasureValueType": measure_value_type
+                    }
+
+                    measure_value = None
+                    if "random_options" in measure_template:
+                        measure_value = random.choice(measure_template["random_options"])
+                    else:
+                        if current_date == start_date:
+                            if measure_value_type == "DOUBLE":
+                                measure_value = random.uniform(min_value, max_value)
+                            elif measure_value_type == "VARCHAR":
+                                measure_value = self._generate_random_string(varchar_length)
+                            elif measure_value_type == "BIGINT":
+                                measure_value = int(random.uniform(min_value, max_value))
+                            elif measure_value_type == "BOOLEAN":
+                                measure_value = random.choice([True, False])
+                            else:
+                                raise Exception("Measure value type not recognized")
+                        else:
+                            if measure_value_type == "DOUBLE":
+                                measure_value = max(min_value, min(entity["latest_measures"][measure_name] + random.uniform(-max_variation, max_variation), max_value))
+                            elif measure_value_type == "VARCHAR":
+                                measure_value = self._generate_random_string(varchar_length)
+                            elif measure_value_type == "BIGINT":
+                                measure_value = int(max(min_value, min(entity["latest_measures"][measure_name] + int(random.uniform(-max_variation, max_variation)), max_value)))
+                            elif measure_value_type == "BOOLEAN":
+                                measure_value = random.choice([True, False])
+                            else:
+                                raise Exception("Measure value type not recognized")
+                        
+                    # Store the actual value in the entity for future iteration
+                    entity["latest_measures"][measure_name] = measure_value
+
+                    # Timestream requires that all data be inserted as a string
+                    measure["MeasureValue"] = str(measure_value)
+
+                    measures.append(measure)
+                
+                precision = precision.strip().upper()
+                if precision == "SECONDS":
+                    timestamp = str(int(current_date.timestamp()))
+                if precision == "MICROSECONDS":
+                    timestamp = str(int(current_date.timestamp() * 1_000_000))
+                else:
+                    # Default to millisecond precision
+                    timestamp = str(int(current_date.timestamp() * 1_000))
+
+                record = {
+                    "Dimensions": dimensions,
+                    "Time": timestamp,
+                    "Measures": measures
+                }
+                records.append(record)
+            current_date += reporting_frequency
+        return records
+    
+    def _generate_random_string(self, length: int) -> str:
+        """
+        Generates a random alphanumeric string.
+
+        :param length: The length of the string to generate.
+        :returns: A random string.
+        """
+
+        letters = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"
+        return ''.join(random.choice(letters) for _ in range(length))
+    
+    def generate_dashboard(self, grafana_data_source_name: str, database_name: str, table_name: str) -> dict:
+        """
+        Generates a Grafana dashboard.
+
+        :param grafana_data_source_name: str: The name of an existing Timestream for LiveAnalytics data source in Grafana.
+        :param database_name: str: The Timestream for LiveAnalytics database the panels will use.
+        :param table_name: str: The Timestream for LiveAnalytics table the panels will use.
+        :returns: A Grafana dashboard as a dict.
+        """
+
+        # Each measure will have a panel
+        panels = []
+        for _, measure_template in enumerate(self.measure_templates):
+            measure_name = measure_template['name']
+            query = ""
+            if len(self.measure_templates) > 1:
+                dimensions = ', '.join(f"\"{dimension_template['name']}\"" for dimension_template in self.dimension_templates)
+                query = f'SELECT time, "{measure_name}", {dimensions} FROM "{database_name}"."{table_name}" ORDER BY time ASC'
+            elif len(self.measure_templates) == 1:
+                query = f'SELECT * FROM "{database_name}"."{table_name}" ORDER BY time ASC'
+            
+            panel = {
+                "datasource": grafana_data_source_name,
+                "gridPos": {
+                    "h": 22,
+                    "w": 20,
+                    "x": 0,
+                    "y": 0
+                },
+                "targets": [
+                    {
+                        "datasource": grafana_data_source_name,
+                        "rawQuery": query,
+                    }
+                ],
+                "title": f"{measure_name}",
+                "type": "timeseries"
+            }
+            panels.append(panel)
+
+        dashboard = {
+            "panels": panels,
+            "title": "Amazon Timestream for LiveAnalytics Sample Dashboard"
+        }
+
+        return dashboard
